added code coverage exclusion on temporary package
