+++ dealmate-project/user-service/src/main/java/pl/zzpj/dealmate/userservice/controller/UserController.java
@@ -1,17 +1,13 @@
 package pl.zzpj.dealmate.userservice.controller;
 
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 import pl.zzpj.dealmate.userservice.model.UserEntity;
 import pl.zzpj.dealmate.userservice.payload.request.RegisterRequest;
 import pl.zzpj.dealmate.userservice.service.UserService;
 
-import java.util.Enumeration;
 
 @Slf4j
 @RestController
@@ -22,38 +18,6 @@ public class UserController {
 
     public UserController(UserService userService) {
         this.userService = userService;
-    }
-
-    @GetMapping("authtest")
-    public ResponseEntity<String> authTest(Authentication authentication, HttpServletRequest request) {
-        Enumeration<String> headerNames = request.getHeaderNames();
-        StringBuilder headers = new StringBuilder();
-        while (headerNames.hasMoreElements()) {
-            String headerName = headerNames.nextElement();
-            headers.append(headerName).append(": ").append(request.getHeader(headerName)).append("\n");
-        }
-
-        if (authentication == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .body("Authentication object is null\nHeaders:\n" + headers.toString());
-        }
-
-        try {
-            String className = authentication.getClass().getName();
-
-            if (authentication instanceof JwtAuthenticationToken jwtAuthenticationToken) {
-                String username = authentication.getName();
-                String jwtString = jwtAuthenticationToken.getToken().getTokenValue();
-
-                return ResponseEntity.ok("Success! Hi " + username + ", jwt: " + jwtString);
-            } else {
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                        .body("Authentication is not JwtAuthenticationToken but: " + className);
-            }
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("Error: " + e.getMessage() + "\nHeaders:\n" + headers.toString());
-        }
     }
 
     @PostMapping("/register")
