+++ sonar-project.properties
@@ -21,7 +21,7 @@ sonar.coverage.jacoco.xmlReportPaths=dealmate-project/*/target/site/jacoco/jacoc
 
 # Exclude files from the test coverage analysis
 # ? Should dtos, configs, entities, and enums be excluded from the analysis?
-sonar.coverage.exclusions=dealmate-project/**/*Application.java,dealmate-project/*/src/test/**,dealmate-project/config-server/**,dealmate-project/discovery-server/**,dealmate-project/auth-server/**
+sonar.coverage.exclusions=dealmate-project/**/*Application.java,dealmate-project/**/temporary,dealmate-project/*/src/test/**,dealmate-project/config-server/**,dealmate-project/discovery-server/**,dealmate-project/auth-server/**
 
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
